pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriel",email = "gabrielsilvalima1994@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-debug-toolbar (>=6.2.0,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "gitpython (>=3.1.46,<4.0.0)",
    "pillow (>=12.1.0,<13.0.0)"
]

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
